chore: refactor instance and setup transformer to dedicated files
